Move constants into CordappInfo companion object.
