+++ core/src/main/kotlin/net/corda/core/contracts/ContractAttachment.kt
@@ -1,8 +1,8 @@
 package net.corda.core.contracts
 
 import net.corda.core.KeepForDJVM
+import net.corda.core.internal.cordapp.CordappInfo.Companion.DEFAULT_CORDAPP_VERSION
 import net.corda.core.serialization.CordaSerializable
-import net.corda.core.internal.cordapp.DEFAULT_CORDAPP_VERSION
 import java.security.PublicKey
 
 /**
