+++ core/src/main/kotlin/net/corda/core/internal/cordapp/CordappInfo.kt
@@ -2,26 +2,27 @@ package net.corda.core.internal.cordapp
 
 /** abstract class used to maintain backwards compatibility */
 abstract class CordappInfo(open val name: String, open val vendor: String, open val version: String, open val minimumPlatformVersion: Int, open val targetPlatformVersion: Int) {
+    companion object {
+        /** CorDapp manifest entries */
+        const val CORDAPP_CONTRACT_NAME = "Cordapp-Contract-Name"
+        const val CORDAPP_CONTRACT_VERSION = "Cordapp-Contract-Version"
+        const val CORDAPP_CONTRACT_VENDOR = "Cordapp-Contract-Vendor"
+        const val CORDAPP_CONTRACT_LICENCE = "Cordapp-Contract-Licence"
+
+        const val CORDAPP_WORKFLOW_NAME = "Cordapp-Workflow-Name"
+        const val CORDAPP_WORKFLOW_VERSION = "Cordapp-Workflow-Version"
+        const val CORDAPP_WORKFLOW_VENDOR = "Cordapp-Workflow-Vendor"
+        const val CORDAPP_WORKFLOW_LICENCE = "Cordapp-Workflow-Licence"
+
+        const val TARGET_PLATFORM_VERSION = "Target-Platform-Version"
+        const val MIN_PLATFORM_VERSION = "Min-Platform-Version"
+
+        const val UNKNOWN_VALUE = "Unknown"
+        const val DEFAULT_CORDAPP_VERSION = 1
+    }
     abstract fun description(): String
     abstract fun hasUnknownFields(): Boolean
 }
-
-/** CorDapp manifest entries */
-const val CORDAPP_CONTRACT_NAME = "Cordapp-Contract-Name"
-const val CORDAPP_CONTRACT_VERSION = "Cordapp-Contract-Version"
-const val CORDAPP_CONTRACT_VENDOR = "Cordapp-Contract-Vendor"
-const val CORDAPP_CONTRACT_LICENCE = "Cordapp-Contract-Licence"
-
-const val CORDAPP_WORKFLOW_NAME = "Cordapp-Workflow-Name"
-const val CORDAPP_WORKFLOW_VERSION = "Cordapp-Workflow-Version"
-const val CORDAPP_WORKFLOW_VENDOR = "Cordapp-Workflow-Vendor"
-const val CORDAPP_WORKFLOW_LICENCE = "Cordapp-Workflow-Licence"
-
-const val TARGET_PLATFORM_VERSION = "Target-Platform-Version"
-const val MIN_PLATFORM_VERSION = "Min-Platform-Version"
-
-const val UNKNOWN_VALUE = "Unknown"
-const val DEFAULT_CORDAPP_VERSION = 1
 
 /** new identifiers (from Corda 4) */
