+++ testing/node-driver/src/test/kotlin/net/corda/testing/node/internal/TestCordappsUtilsTest.kt
@@ -1,10 +1,10 @@
 package net.corda.testing.node.internal
 
-import net.corda.core.internal.cordapp.CORDAPP_CONTRACT_NAME
-import net.corda.core.internal.cordapp.CORDAPP_WORKFLOW_NAME
-import net.corda.core.internal.cordapp.TARGET_PLATFORM_VERSION
-import net.corda.core.internal.inputStream
+import net.corda.core.internal.cordapp.CordappInfo.Companion.CORDAPP_CONTRACT_NAME
+import net.corda.core.internal.cordapp.CordappInfo.Companion.CORDAPP_WORKFLOW_NAME
+import net.corda.core.internal.cordapp.CordappInfo.Companion.TARGET_PLATFORM_VERSION
 import net.corda.core.internal.cordapp.get
+import net.corda.core.internal.inputStream
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Rule
 import org.junit.Test
