+++ testing/test-utils/src/main/kotlin/net/corda/testing/core/internal/ContractJarTestUtils.kt
@@ -1,18 +1,21 @@
 package net.corda.testing.core.internal
 
-import net.corda.core.internal.cordapp.CORDAPP_CONTRACT_VERSION
-import net.corda.testing.core.internal.JarSignatureTestUtils.addManifest
-import net.corda.testing.core.internal.JarSignatureTestUtils.createJar
-import net.corda.testing.core.internal.JarSignatureTestUtils.generateKey
-import net.corda.testing.core.internal.JarSignatureTestUtils.signJar
+import net.corda.core.internal.cordapp.CordappInfo.Companion.CORDAPP_CONTRACT_VERSION
 import net.corda.core.internal.delete
 import net.corda.core.internal.div
 import net.corda.core.internal.toPath
 import net.corda.testing.core.ALICE_NAME
+import net.corda.testing.core.internal.JarSignatureTestUtils.addManifest
+import net.corda.testing.core.internal.JarSignatureTestUtils.createJar
+import net.corda.testing.core.internal.JarSignatureTestUtils.generateKey
+import net.corda.testing.core.internal.JarSignatureTestUtils.signJar
 import java.io.OutputStream
 import java.net.URI
 import java.net.URL
-import java.nio.file.*
+import java.nio.file.Files
+import java.nio.file.Path
+import java.nio.file.Paths
+import java.nio.file.StandardCopyOption
 import java.security.PublicKey
 import java.util.jar.Attributes
 import java.util.jar.JarEntry
